Revised CovJSON response handler requirements
Previous requirements were poorly defined and did not properly reflect the requirements of the project. These updated requirements should be sufficient for the remainder of the project with little to no revisions. Struck out "Original Requirements" section and added an "Updated Requirements" section with the revised requirements. Made some other minor changes as well.